user a better way to handler wait on close
